third_party/WebKit/LayoutTests/inspector-protocol/dom-snapshot/dom-snapshot-getSnapshot.html

<html>
<head>
<style>
@font-face {
  font-family: 'ahem';
  src: url(../../resources/Ahem.ttf);
}
</style>
<script type="text/javascript" src="../../http/tests/inspector-protocol/resources/inspector-protocol-test.js"></script>
<script>

async function test() {
    function stabilize(key, value) {
        var unstableKeys = ["documentURL", "baseURL", "frameId"];
        if (unstableKeys.indexOf(key) !== -1)
            return "<" + typeof(value) + ">";
        return value;
    }
  
    var whitelist = ["transform", "transform-origin", "height", "width", "display", "outline-color", "color"];
    var response = await InspectorTest.sendCommandOrDie("DOMSnapshot.getSnapshot", {"computedStyleWhitelist": whitelist});
    InspectorTest.log(JSON.stringify(response, stabilize, 2));
    InspectorTest.completeTest();
}

</script>
<template id="shadow-template">
<style>
:host {
    color: red;
}
</style>
<div style="font-family: ahem;"><h1>Hi!</h1></div>
</template>
</head>
<body class="body-class">
<div style="font-family: ahem;">
    <div class="class1">Some Text</div> And More Text
    <div style="display:inline-block; width: 200px">
        <p>
        Lorem ipsum dolor sit amet, consectetur adipiscing elit. Pellentesque sit amet est sem.
        Aenean ut neque volutpat, posuere odio at, mollis nibh. Aenean sodales nulla et
        ligula efficitur sollicitudin blandit sed lectus. Duis orci enim, sodales ac lectus sed,
        hendrerit efficitur est. Quisque gravida facilisis viverra.
        </p>
        <ul class="class3">
            <li class="class4"></li>
            <span>Lets have a span</span>
        </ul>
    </div>
    <div style="transform: rotateZ(90deg); width: 200px">Rotated text!</div>
    <iframe src="../dom/resources/simple-iframe.html" width="400" height="200"></iframe>
    <div id="shadow-host"></div>
    <script type="text/javascript">
        var host = document.querySelector("#shadow-host").createShadowRoot();
        var template = document.querySelector("#shadow-template");
        host.appendChild(template.content);
        template.remove();
        window.onload = runTest;
    </script>
</div>
</body>
</html>
